Fix: int types

+++ src/mympd_api/queue.h
@@ -12,7 +12,7 @@
 
 bool mympd_api_queue_save(struct t_partition_state *partition_state, sds name, sds mode, sds *error);
 sds mympd_api_queue_list(struct t_partition_state *partition_state, sds buffer, long request_id,
-        long offset, long limit, const struct t_tags *tagcols);
+        unsigned offset, unsigned limit, const struct t_tags *tagcols);
 sds mympd_api_queue_crop(struct t_partition_state *partition_state, sds buffer, enum mympd_cmd_ids cmd_id,
         long request_id, bool or_clear);
 sds mympd_api_queue_search(struct t_partition_state *partition_state, sds buffer, long request_id,
